+++ src/Ninja/G1CP/Content/Fixes/Gamesave/fix051_StonehengeCryptChest.d
@@ -2,8 +2,8 @@
  * #51 Chest in stonehenge crypt behind wall
  */
 func int G1CP_051_StonehengeCryptChest() {
-    var int vobPtr; vobPtr = G1CP_FindVobByPosF(-35498.5625, 2331.01416, -13830.9131);
-    if (Hlp_Is_oCMobContainer(vobPtr)) { // Make sure we found a (the) chest
+    var int vobPtr; vobPtr = G1CP_FindVobByPosF(-35498.5625, 2331.01416, -13830.9131, Hlp_Is_oCMobContainer);
+    if (vobPtr) {
         G1CP_MoveVobToPosF(vobPtr, -35311.8100, 2281.0100, -14096.8200);
         return TRUE;
     };
@@ -20,8 +20,8 @@ func int G1CP_051_StonehengeCryptChestRevert() {
     };
 
     // Search the VOB again
-    var int vobPtr; vobPtr = G1CP_FindVobByPosF(-35311.8100, 2281.0100, -14096.8200);
-    if (Hlp_Is_oCMobContainer(vobPtr)) { // Make sure we found a (the) chest
+    var int vobPtr; vobPtr = G1CP_FindVobByPosF(-35311.8100, 2281.0100, -14096.8200, Hlp_Is_oCMobContainer);
+    if (vobPtr) {
         G1CP_MoveVobToPosF(vobPtr, -35498.5625, 2331.01416, -13830.9131);
         return TRUE;
     };