+++ src/Ninja/G1CP/Content/Fixes/Gamesave/fix052_UseWithItemNcGrindstone.d
@@ -12,8 +12,8 @@ func int G1CP_052_UseWithItemNcGrindstone() {
     };
 
     // Search the VOB
-    var int vobPtr; vobPtr = G1CP_FindVobByPosF(-58212.9141, 3233.08716, 7490.75928);
-    if (Hlp_Is_oCMobInter(vobPtr)) {
+    var int vobPtr; vobPtr = G1CP_FindVobByPosF(-58212.9141, 3233.08716, 7490.75928, Hlp_Is_oCMobInter);
+    if (vobPtr) {
         var oCMobInter mob; mob  = _^(vobPtr);
         if (Hlp_StrCmp(mob.sceme, "BSSHARP"))
         && (Hlp_StrCmp(mob.useWithItem, "")) {
@@ -34,8 +34,8 @@ func int G1CP_052_UseWithItemNcGrindstoneRevert() {
     };
 
     // Search the VOB again
-    var int vobPtr; vobPtr = G1CP_FindVobByPosF(-58212.9141, 3233.08716, 7490.75928);
-    if (Hlp_Is_oCMobInter(vobPtr)) {
+    var int vobPtr; vobPtr = G1CP_FindVobByPosF(-58212.9141, 3233.08716, 7490.75928, Hlp_Is_oCMobInter);
+    if (vobPtr) {
         var oCMobInter mob; mob  = _^(vobPtr);
         if (Hlp_StrCmp(mob.sceme, "BSSHARP"))
         && (Hlp_StrCmp(mob.useWithItem, "ITMISWORDBLADE")) {