+++ src/Ninja/G1CP/Content/Fixes/Session/fix122_CavalornDailyRoutine.d
@@ -64,10 +64,11 @@ func int G1CP_122_CavalornDailyRoutine() {
     // Check if any matches are left
     if (MEM_ArraySize(sleepMatches) > 0) && (MEM_ArraySize(standMatches) > 0) {
         // Get symbol indices of the strings
+        var MEMINT_HelperClass ref;
         const string wpSleep = "OW_SAWHUT_SLEEP_01";
         const string wpStand = "OW_CAVALORN_01";
-        var int wpSleepSymbId; wpSleepSymbId = MEM_GetSymbolIndex("G1CP_122_CavalornDailyRoutine.wpSleep");
-        var int wpStandSymbId; wpStandSymbId = MEM_GetSymbolIndex("G1CP_122_CavalornDailyRoutine.wpStand");
+        var int wpSleepSymbId; wpSleepSymbId = ref + 1;
+        var int wpStandSymbId; wpStandSymbId = ref + 2;
 
         // Switch the waypoints in all remaining occurrences
         repeat(i, MEM_ArraySize(sleepMatches));