+++ src/Ninja/G1CP/Content/Tests/test144.d
@@ -6,5 +6,5 @@ func int G1CP_Test_144() {
     var C_Item itm; itm = G1CP_Testsuite_CreateItem("EBR_ARMOR_H");
     G1CP_Testsuite_CheckPassed();
 
-    return G1CP_Testsuite_InspectItemVariable(itm, "name", "Gomez' Rüstung");
+    return G1CP_Testsuite_InspectItemString(itm, "name", "Gomez' Rüstung");
 };