+++ src/Ninja/G1CP/Content/Tests/test173.d
@@ -6,5 +6,5 @@ func int G1CP_Test_173() {
     var C_Item itm; itm = G1CP_Testsuite_CreateItem("ItKe_Gomez_01");
     G1CP_Testsuite_CheckPassed();
 
-    return G1CP_Testsuite_InspectItemVariable(itm, "text[0]", "Öffnet Gomez' Truhen");
+    return G1CP_Testsuite_InspectItemString(itm, "text[0]", "Öffnet Gomez' Truhen");
 };