+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## v1.2.0 (TBA)
+## v1.3.0 (TBA)
+### General
+* Fix [#278](https://g1cp.org/issues/278): The standard voice subtitles of female NPCs when approaching them now match the audio.
+
+### Story
+* Fix [#57](https://g1cp.org/issues/57): The wall texture of the castle of the Old Camp near Gravo does not flicker anymore.
+* Fix [#116](https://g1cp.org/issues/116): There are no longer lens flares in Jacko's swampweed cave.
+
+## [v1.2.0](https://g1cp.org/releases/tag/v1.2.0) (2021-10-31)
 ### General
 * Fix [#58](https://g1cp.org/issues/58): Falling can no longer be interrupted by fight actions mid-air.
 * Fix [#194](https://g1cp.org/issues/194): NPCs now correctly pick up the weapon of their defeated opponent.